✨ (App.jsx, DNSLookupTool.jsx): refactor DNS lookup tool by renaming component to DNSLookupShadcn and removing the old DNSLookupTool implementation to streamline the codebase.
